management_system/requirements.txt
django == 4.2.4
pillow ==10.0.1
psycopg2-binary == 2.9.7
python-decouple == 3.8
djangorestframework 
djangorestframework-jwt
